+++ opraas_core/src/infra/ethereum/mod.rs
@@ -1,10 +1,9 @@
-pub mod eth_rpc;
-pub mod eth_rpc_json;
+pub mod rpc;
+pub mod rpc_json;
+pub use rpc::*;
+pub use rpc_json::*;
 
 pub mod node;
 pub mod node_geth;
-
-pub use eth_rpc::*;
-pub use eth_rpc_json::*;
 pub use node::*;
 pub use node_geth::*;
