+++ opraas_core/src/infra/ethereum/node_geth.rs
@@ -1,7 +1,6 @@
 use crate::system::execute_command;
 use serde_json::json;
 use std::{process::Command, thread, time};
-
 use super::{EthRpc, JsonRpc, TTestnetNode};
 
 const DOCKER_IMAGE: &str = "ethereum/client-go:v1.13.4";
